Move only the variables and make them constants in the platform specific file instead of the entire function
